implement objExplorer srna
