+++ src/components/ui-components/web/objectListExplorer/table.module.css
@@ -1,0 +1,4 @@
+
+.cell:hover{
+    background-color: #cadce7;
+}
